limit user permissions   - user can update, delete his own profile (or admin role)   - user can update, delete his own travel (or admin role)   - user can see public travels (admin see all travels)   -  user can see a specific public travel   -  creating destination create relationship with categories   -  can update categories of a destination   -  list destination by category controller created   - fix relationship between travel and categories (O to M)

+++ api/controllers/user.controller.js
@@ -33,6 +33,10 @@ const getOneUser = async (req, res) => {
 
 const updateUser = async (req, res) => {
   try {
+    if (req.params.id !== res.locals.user.id || res.locals.user.roles !== 'admin') {
+			return res.status(500).send('You are not authorized to access this resource')
+    }
+
     if(req.body.password) { req.body.password = bcrypt.hashSync(req.body.password, 10) }
 
 		const [userExist, user] = await User.update(req.body, {
@@ -52,6 +56,10 @@ const updateUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
   try {
+    if (req.params.id !== res.locals.user.id || res.locals.user.roles !== 'admin') {
+			return res.status(500).send('You are not authorized to access this resource')
+    }
+    
     const user = await User.destroy({
 			where: { id: req.params.id },
 		})
@@ -67,7 +75,11 @@ const deleteUser = async (req, res) => {
 }
 
 const showProfile = async (req, res) => {
-  return res.json(res.locals.user)
+  try {
+    return res.json( res.locals.user )
+  } catch (err) {
+    return res.status(404).send('Error: User not found') 
+  }
 }
 
 module.exports = { createUser, getUsers, getOneUser, updateUser, deleteUser, showProfile }
